requirements.txt
setuptools
Django==2.2
psycopg2==2.8.6
django-modeltranslation==0.13
django-common-helpers==0.9.2
django-simple-captcha==0.5.9
django-cron==0.5.1
django-haystack==2.8.1
django-summernote==0.8.11.1
django-transmeta==0.7.3
httplib2==0.12.0
Pillow==5.3.0
pytz==2018.7
django-filter==2.0.0
django-material==1.4.3
django-viewflow==1.4.2
djangorestframework==3.9.0
markdown==3.0.1
django-bootstrap-form==3.4
django-cors-headers==2.4.0
django-modeltranslation==0.13
requests==2.21.0
psycopg2-binary
uwsgi
graphene-django==2.9.1
wheel
